requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
plotly
uvicorn
fastapi
pymongo
bcrypt
passlib[bcrypt]
python-dotenv
python-jose[cryptography]
